refactor(deployment): move component class functions to singletons in util files
