+++ lib/manager/components/DeploymentConfirmModal.js
@@ -4,7 +4,7 @@ import {Map, TileLayer, Rectangle} from 'react-leaflet'
 
 import {getComponentMessages, getMessage} from '../../common/util/config'
 import {defaultTileURL} from '../../common/util/maps'
-import {versionHasExpired} from '../util'
+import {getFeedNames, versionHasExpired} from '../util/version'
 import polygon from 'turf-polygon'
 import area from '@turf/area'
 
@@ -28,8 +28,6 @@ export default class DeploymentConfirmModal extends Component {
   }
 
   _onClose = () => this.props.onClose()
-
-  _getFeedNames = (versions) => versions.map(v => v.feedSource.name).join(', ')
 
   render () {
     const {Body, Footer, Header, Title} = Modal
@@ -65,7 +63,7 @@ export default class DeploymentConfirmModal extends Component {
             </li>
             <li>
               <Glyphicon glyph='list' />{' '}Deploying {deployment.feedVersions.length} feeds:{' '}
-              {this._getFeedNames(deployment.feedVersions)}
+              {getFeedNames(deployment.feedVersions)}
             </li>
           </ul>
           <h3>OpenTripPlanner Settings</h3>
@@ -103,7 +101,7 @@ export default class DeploymentConfirmModal extends Component {
                 : null
           }
           {expiredFeeds.length > 0 &&
-            <Alert label={`${getMessage(this.messages, 'alert.expiredFeeds')}: ${this._getFeedNames(expiredFeeds)}`} />
+            <Alert label={`${getMessage(this.messages, 'alert.expiredFeeds')}: ${getFeedNames(expiredFeeds)}`} />
           }
           {isDeployed && oldDeployment // Show succces alert.
             ? <Alert success label={getMessage(this.messages, 'alert.success')} />
