+++ lib/manager/components/DeploymentSettings.js
@@ -185,7 +185,6 @@ class DeploymentSettings extends Component {
     .indexOf(false) === -1
 
   render () {
-    console.log(this.state)
     const updaters = objectPath.get(this.state, 'routerConfig.updaters') || []
     const {otpServers} = this.state
     const messages = getComponentMessages('ProjectSettings')
