+++ lib/manager/components/DeploymentVersionsTable.js
@@ -3,7 +3,7 @@ import {Table, Button, Badge, Glyphicon, ButtonToolbar} from 'react-bootstrap'
 import { Link } from 'react-router'
 
 import { getComponentMessages, getMessage } from '../../common/util/config'
-import {versionHasExpired, versionHasNotBegun} from '../util'
+import {versionHasExpired, versionHasNotBegun} from '../util/version'
 import {formatTimestamp, fromNow} from '../../common/util/date-time'
 
 export default class DeploymentVersionsTable extends Component {
