+++ lib/manager/components/DeploymentViewer.js
@@ -14,6 +14,7 @@ import DeploymentConfirmModal from './DeploymentConfirmModal'
 import DeploymentPreviewButton from './DeploymentPreviewButton'
 import DeploymentVersionsTable from './DeploymentVersionsTable'
 import {formatTimestamp, fromNow} from '../../common/util/date-time'
+import {isValidJSON} from '../../common/util/json'
 
 const BOUNDS_LIMIT = 10 // Limit for the decimal degrees span
 
@@ -49,6 +50,12 @@ export default class DeploymentViewer extends Component {
     this.resetMap()
   }
 
+  /**
+   * When the component remounts after an initial mount, the map needs to be
+   * reset using Leaflet.invalidateSize so that the map tiles do not "grey out."
+   *
+   * Also, this re-fits the bounds to the deployment bounds.
+   */
   resetMap () {
     setTimeout(() => {
       if (this.refs.map) {
@@ -345,19 +352,10 @@ class CustomConfig extends Component {
   _onChangeConfig = ({target}) =>
     this.setState({[this.props.name]: target.value})
 
-  _isValidJSON = (str) => {
-    try {
-      JSON.parse(str)
-    } catch (e) {
-      return false
-    }
-    return true
-  }
-
   _onSaveConfig = () => {
     const {deployment, name, updateDeployment} = this.props
     const value = this.state[name]
-    if (!this._isValidJSON(value)) return window.alert('Must provide valid JSON string.')
+    if (!isValidJSON(value)) return window.alert('Must provide valid JSON string.')
     else {
       updateDeployment(deployment, {[name]: value})
       this.setState({[name]: undefined})
@@ -368,7 +366,7 @@ class CustomConfig extends Component {
     const {deployment, name, label} = this.props
     const useCustom = deployment[name] !== null
     const value = this.state[name] || deployment[name]
-    const validJSON = this._isValidJSON(value)
+    const validJSON = isValidJSON(value)
     return (
       <div>
         <h5>{label} configuration</h5>
