+++ lib/manager/util/index.js
@@ -1,7 +1,5 @@
 // @flow
 
-import moment from 'moment'
-
 import type {Feed, FeedVersion, Project, User} from '../../types'
 
 export function findProjectByFeedSource (
@@ -34,14 +32,6 @@ export function getChartPeriod (maxValue: number): number {
           : maxValue > 20
             ? 10
             : 3
-}
-
-export function versionHasExpired (version: FeedVersion): boolean {
-  return moment(version.validationResult.endDate).isBefore(moment())
-}
-
-export function versionHasNotBegun (version: FeedVersion): boolean {
-  return moment(version.validationResult.startDate).isAfter(moment())
 }
 
 export function isEditingDisabled (
