+++ lib/types.js
@@ -138,6 +138,7 @@ export type Fare = {
 
 export type Feed = {
   id: string,
+  name: string,
   latestValidation?: {
     bounds: Bounds
   }
@@ -145,7 +146,10 @@ export type Feed = {
 
 export type FeedVersion = {
   id: string,
+  name: string,
+  feedLoadResult: any,
   feedSourceId: string,
+  feedSource: Feed,
   validationResult: {
     endDate: string,
     startDate: string
